rl_service/requirements.txt
Flask
Flask-CORS
numpy
python-dotenv 